requirements.txt
Flask
Pillow
Flask-Cors
Celery[Redis]